# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Наша CRM",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/category": {
            "get": {
                "tags": [
                    "Категории"
                ],
                "summary": "Получить список Категорий",
                "description": "Вернет список Категорий",
                "operationId": "6dd4a3f89346cf8364fb77e36432b31e",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "categories": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": ""
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": ""
                                                    },
                                                    "limit": {
                                                        "type": "integer",
                                                        "example": ""
                                                    }
                                                },
                                                "type": "object"
                                            },
                                            "example": [
                                                {
                                                    "id": 1,
                                                    "name": "Fanger",
                                                    "limit": 12312
                                                },
                                                {
                                                    "id": "",
                                                    "name": "",
                                                    "limit": ""
                                                }
                                            ]
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/category/{category_id}": {
            "get": {
                "tags": [
                    "Категории"
                ],
                "summary": "Получить категорию по id",
                "description": "Вернет категорию по id",
                "operationId": "9773d99b8e5e81e25575f5e0d9737ab5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID категории",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "example": "1"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "test"
                                        },
                                        "limit": {
                                            "type": "integer",
                                            "example": 123
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/category/add": {
            "post": {
                "tags": [
                    "Категории"
                ],
                "summary": "Добавить категорию",
                "description": "Нужно передать все поля",
                "operationId": "6baaa1e917d709ce91e18a335435b052",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "limit"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "test"
                                    },
                                    "limit": {
                                        "type": "integer",
                                        "example": 6521
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "422": {
                        "description": "Вернет ошибку валидации, если поля не валидны или какие-то не отправлены"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/category/update/{category_id}": {
            "put": {
                "tags": [
                    "Категории"
                ],
                "summary": "Обновить Категрию",
                "description": "Нужно передать все поля",
                "operationId": "f5816d097ab76d537df745feb1e981a1",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "limit"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "test"
                                    },
                                    "limit": {
                                        "type": "integer",
                                        "example": 6521
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "422": {
                        "description": "Вернет ошибку валидации, если поля не валидны или какие-то не отправлены"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/category/delete/{category_id}": {
            "delete": {
                "tags": [
                    "Категории"
                ],
                "summary": "Удалить категорию по id",
                "description": "Вернет булевое значение",
                "operationId": "d5e0a6de0c9a4713f9239dad65563ba2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID категории",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "example": "1"
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
...
    }
}